+++ spec/models/admin_dashboard_report_spec.rb
@@ -1,6 +1,23 @@
 # frozen_string_literal: true
 
 RSpec.describe AdminDashboardReport do
+  let(:another_provider) do
+    Class.new(AdminDashboard::Reports::SourceProvider) { def self.source_name = "another_source" }
+  end
+
+  before do
+    DiscoursePluginRegistry.register_admin_dashboard_report_source(
+      another_provider,
+      Plugin::Instance.new,
+    )
+  end
+
+  after do
+    DiscoursePluginRegistry._raw_admin_dashboard_report_sources.reject! do |entry|
+      entry[:value] == another_provider
+    end
+  end
+
   describe "validations" do
     it "requires source and identifier" do
       record = described_class.new
@@ -19,8 +36,14 @@ RSpec.describe AdminDashboardReport do
     it "allows the same identifier under a different source" do
       described_class.create!(source: "core_report", identifier: "signups")
 
-      other_source = described_class.new(source: "data_explorer_query", identifier: "signups")
+      other_source = described_class.new(source: "another_source", identifier: "signups")
       expect(other_source).to be_valid
+    end
+
+    it "rejects sources without a registered provider" do
+      record = described_class.new(source: "totally_unregistered", identifier: "x")
+      expect(record).not_to be_valid
+      expect(record.errors.attribute_names).to include(:source)
     end
   end
